+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/maven/MojoVerify.java
@@ -33,6 +33,9 @@ public class MojoVerify extends AbstractMojo {
     @Parameter(property = "reportOutputFolder", required = false)
     protected String reportOutputFolder;
 
+    @Parameter(property = "reportOutputFormat", defaultValue = "html", required = false)
+    protected String reportOutputFormat;
+
     private void addIfPresent( Properties params, String key, String value ) {
         if ( value != null ) {
             this.getLog().info( String.format( "setting verifier parameter : %s -> %s", key, value ) );
@@ -54,6 +57,7 @@ public class MojoVerify extends AbstractMojo {
         this.addIfPresent( params, FreeMarkerTemplateSyntaxVerifier.PARAM_TEMPLATE_FILE_PATTERN, this.templateFilePattern );
         this.addIfPresent( params, FreeMarkerTemplateSyntaxVerifier.PARAM_GENERATE_REPORT, String.valueOf( this.generateReport ) );
         this.addIfPresent( params, FreeMarkerTemplateSyntaxVerifier.PARAM_REPORT_OUTPUT_FOLDER, this.reportOutputFolder );
+        this.addIfPresent( params, FreeMarkerTemplateSyntaxVerifier.PARAM_REPORT_OUTPUT_FORMAT, this.reportOutputFormat );
         FreeMarkerTemplateSyntaxVerifier verifier = new FreeMarkerTemplateSyntaxVerifier();
         VerifyTemplateOutput output = verifier.createConfigurationAndVerify( baseFolder, params );
         this.getLog().info( String.format( "verify output resultCode %s, checked templates : %s", output.getResultCode(), output.getInfos().size() ) );